simplify filter_options
